_backend/requirements.txt
beautifulsoup4==4.8.2
bottle==0.12.18
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.8
jieba3k==0.35.1
lxml==4.5.0
newspaper3k==0.2.8
nltk==3.4.5
Pillow==7.0.0
pip==9.0.1
pymongo==3.10.1
python-dateutil==2.8.1
PyYAML==5.3
requests==2.22.0
requests-file==1.4.3
setuptools==28.8.0
six==1.14.0
soupsieve==1.9.5
tinysegmenter==0.3
tldextract==2.2.2
urllib3==1.25.8
